Added a barrier so that context worker threads wait until their locality has been initialized before they intialize themselves.
This allows me to register initialization callbacks for kthread local data.

Also did some de-indentation in kthread.c and ctx.c.

+++ include/sync/locks.h
@@ -17,11 +17,10 @@
 #ifndef HPX_SYNC_LOCKS_H_
 #define HPX_SYNC_LOCKS_H_
 
-#include <stdint.h>
 #include "sync.h"
 
 struct tatas_lock {
-    uintptr_t lock;
+    int lock;
 };
 
 typedef struct tatas_lock tatas_lock_t;
